requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
pillow
python-multipart
